requirements.txt
gTTS
pyaudio
playsound==1.2.2
deep-translator
SpeechRecognition
google-transliteration-api